Add CAD for complete Sandbox + some parts
