+++ assembly/readme.md
@@ -1,18 +1,23 @@
 # Assembly
 
+Besides each assembly section, the CAD of the Sandbox can be used as a guide for assembling the Sandbox.
+
+- FreeCAD file: [./cad/freecad/tectoh.FCStd](./cad/freecad/tectoh.FCStd)
+
+
 ## CAD
 
 Parts to 3D print and CAD of the system 
 
-[CAD](./cad/readme.md)
+[CAD](./cad/.)
 
 ## Mechanical assembly instructions
 
-[Mechanical assembly](./mechanical/readme.md)
+[Mechanical assembly](./mechanical/.)
 
 ## Electronic schematic
 
-[Electronics assembly](./electronics/readme.md)
+[Electronics assembly](./electronics/.)
