+++ optional/readme.md
@@ -1,4 +1,4 @@
 # Optional components
 
-[High-resolution linear magnetic sensor AS5311](./as5311_magn_sens/readme.md)
+[High-resolution linear magnetic sensor AS5311](./as5311_magn_sens/.)
